Fix TotalByAccount response

+++ src/main/java/banky/webservices/api/evolution/responses/TotalByAccountResponse.java
@@ -1,5 +1,7 @@
 package banky.webservices.api.evolution.responses;
 
+import java.math.BigDecimal;
+
 /**
  * Represents the total amounts for a single account in the evolution report
  * Corresponds to the TotalByAccount type in the frontend
@@ -8,7 +10,7 @@ public record TotalByAccountResponse(
     String id,
     String shortName,
     String name,
-    double total,
-    double gainLoss,
-    double gainLossPercentage
+    BigDecimal total,
+    BigDecimal gainLoss,
+    BigDecimal gainLossPercentage
 ) {}
